extract MessageDeleted to separate component, use smartRender util
